Feedback first round applied
