+++ src/pages/RelayerDashboard/components/Information/index.tsx
@@ -1,6 +1,7 @@
 import { Tabs } from "src/components/organisms";
-import Overview from "./Overview/index";
+import Overview from "./Overview";
 import RawData from "./RawData";
+import Advanced from "./Advanced";
 import { DeliveryLifecycleRecord } from "src/pages/RelayerDashboard/utils/VaaUtils";
 import { useQueries, useQuery } from "react-query";
 import { parseVaa } from "@certusone/wormhole-sdk";
@@ -13,31 +14,28 @@ interface Props {
 }
 
 const Information = ({ lifecycleRecords }: Props) => {
-  const records = lifecycleRecords.filter(record => !!record.vaa);
-  const navigate = useNavigate();
+  // const txsData: any = useQueries(
+  //   lifecycleRecords.map((record, idx) => ({
+  //     queryKey: ["getTransactions", record.vaa],
+  //     queryFn: () => {
+  //       const parsedVaa = parseVaa(record.vaa);
 
-  const txsData: any = useQueries(
-    records.map((record, idx) => ({
-      queryKey: ["getTransactions", record.vaa],
-      queryFn: () => {
-        const parsedVaa = parseVaa(record.vaa);
+  //       const chainId = parsedVaa.emitterChain;
+  //       const emitter = Buffer.from(parsedVaa.emitterAddress).toString("hex");
+  //       const seq = Number(parsedVaa.sequence);
 
-        const chainId = parsedVaa.emitterChain;
-        const emitter = Buffer.from(parsedVaa.emitterAddress).toString("hex");
-        const seq = Number(parsedVaa.sequence);
-
-        return getClient().search.getTransactions({ chainId, emitter, seq });
-      },
-      staleTime: Infinity,
-      onSuccess: (data: GetTransactionsOutput) => {
-        console.log("success getting tx!", idx, data);
-      },
-      onError: (err: unknown) => {
-        console.log("error getting tx", idx, err);
-        // navigate(`/search-not-found/${txHash}`)
-      },
-    })),
-  );
+  //       return getClient().search.getTransactions({ chainId, emitter, seq });
+  //     },
+  //     staleTime: Infinity,
+  //     onSuccess: (data: GetTransactionsOutput) => {
+  //       console.log("success getting tx!", idx, data);
+  //     },
+  //     onError: (err: unknown) => {
+  //       console.log("error getting tx", idx, err);
+  //       // navigate(`/search-not-found/${txHash}`)
+  //     },
+  //   })),
+  // );
 
   // const { payload } = (txData as GetTransactionsOutput) || {};
   // const { payloadType } = payload || {};
@@ -50,13 +48,16 @@ const Information = ({ lifecycleRecords }: Props) => {
       }}
     >
       <Tabs
-        headers={["OVERVIEW", "RAW DATA"]}
+        headers={["OVERVIEW", "ADVANCED", "RAW DATA"]}
         contents={[
           <>
             <Overview
-              txsData={txsData as GetTransactionsOutput[]}
+              // txsData={txsData as GetTransactionsOutput[]}
               lifecycleRecords={lifecycleRecords}
             />
+          </>,
+          <>
+            <Advanced lifecycleRecords={lifecycleRecords} />
           </>,
           <>
             <RawData lifecycleRecords={lifecycleRecords} />
@@ -68,3 +69,14 @@ const Information = ({ lifecycleRecords }: Props) => {
 };
 
 export { Information };
+
+// put the sender address a la derecha de contract address
+
+// if delivery provider matches default one, no need to show it
+// if delivery provider === refund provider ==> not show refund provider
+
+// target address should be the leftiest thing on overview
+
+// binance smart chain --> BNB chain
+
+// parsedVaa.timestamp --> timestamp de la tx mejor.
