+++ src/pages/RelayerDashboard/utils/environment.ts
@@ -59,6 +59,8 @@ export type ChainInfo = {
   rpcUrl: string;
 };
 
+export const tiltDefaultDeliveryProviderContractAddress =
+  "0x1ef9e15c3bbf0555860b5009B51722027134d53a";
 export const tiltEnv: Environment = {
   network: "DEVNET",
   chainInfos: [
@@ -70,7 +72,7 @@ export const tiltEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 1700,
       relayerContractAddress: "0x53855d4b64E9A3CF59A84bc768adA716B5536BC5",
-      defaultDeliveryProviderContractAddress: "0x1ef9e15c3bbf0555860b5009B51722027134d53a",
+      defaultDeliveryProviderContractAddress: tiltDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: "0xC89Ce4735882C9F0f0FE26686c53074E09B0D550",
       mockIntegrationAddress: "0x0eb0dD3aa41bD15C706BC09bC03C002b7B85aeAC",
       rpcUrl: DEVNET_RPCS.ethereum || "",
@@ -83,7 +85,7 @@ export const tiltEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 300,
       relayerContractAddress: "0x53855d4b64E9A3CF59A84bc768adA716B5536BC5",
-      defaultDeliveryProviderContractAddress: "0x1ef9e15c3bbf0555860b5009B51722027134d53a",
+      defaultDeliveryProviderContractAddress: tiltDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: "0xC89Ce4735882C9F0f0FE26686c53074E09B0D550",
       mockIntegrationAddress: "0x0eb0dD3aa41bD15C706BC09bC03C002b7B85aeAC",
       rpcUrl: DEVNET_RPCS.bsc || "",
@@ -92,6 +94,8 @@ export const tiltEnv: Environment = {
   guardianRpcs: ["http://localhost:7071"],
 };
 
+export const testnetDefaultDeliveryProviderContractAddress =
+  "0x60a86b97a7596eBFd25fb769053894ed0D9A8366";
 export const testnetEnv: Environment = {
   network: "TESTNET",
   chainInfos: [
@@ -103,7 +107,7 @@ export const testnetEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 300,
       relayerContractAddress: "0x80aC94316391752A193C1c47E27D382b507c93F3",
-      defaultDeliveryProviderContractAddress: "0x60a86b97a7596eBFd25fb769053894ed0D9A8366",
+      defaultDeliveryProviderContractAddress: testnetDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: "0x68605AD7b15c732a30b1BbC62BE8F2A509D74b4D",
       mockIntegrationAddress: "0xb6A04D6672F005787147472Be20d39741929Aa03",
       rpcUrl: TESTNET_RPCS.bsc || "",
@@ -116,7 +120,7 @@ export const testnetEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 0.66,
       relayerContractAddress: "0x0591C25ebd0580E0d4F27A82Fc2e24E7489CB5e0",
-      defaultDeliveryProviderContractAddress: "0x60a86b97a7596eBFd25fb769053894ed0D9A8366",
+      defaultDeliveryProviderContractAddress: testnetDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: "0x0CBE91CF822c73C2315FB05100C2F714765d5c20",
       mockIntegrationAddress: "0x3bF0c43d88541BBCF92bE508ec41e540FbF28C56",
       rpcUrl: TESTNET_RPCS.polygon || "",
@@ -129,7 +133,7 @@ export const testnetEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 12,
       relayerContractAddress: "0xA3cF45939bD6260bcFe3D66bc73d60f19e49a8BB",
-      defaultDeliveryProviderContractAddress: "0x60a86b97a7596eBFd25fb769053894ed0D9A8366",
+      defaultDeliveryProviderContractAddress: testnetDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: "0x7bbcE28e64B3F8b84d876Ab298393c38ad7aac4C",
       mockIntegrationAddress: "0x5E52f3eB0774E5e5f37760BD3Fca64951D8F74Ae",
       rpcUrl: TESTNET_RPCS.avalanche || "",
@@ -142,7 +146,7 @@ export const testnetEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 0.51,
       relayerContractAddress: "0x306B68267Deb7c5DfCDa3619E22E9Ca39C374f84",
-      defaultDeliveryProviderContractAddress: "0x60a86b97a7596eBFd25fb769053894ed0D9A8366",
+      defaultDeliveryProviderContractAddress: testnetDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: "0x88505117CA88e7dd2eC6EA1E13f0948db2D50D56",
       mockIntegrationAddress: "0x7f1d8E809aBB3F6Dc9B90F0131C3E8308046E190",
       rpcUrl: TESTNET_RPCS.celo || "",
@@ -155,7 +159,7 @@ export const testnetEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 0.26,
       relayerContractAddress: "0x0591C25ebd0580E0d4F27A82Fc2e24E7489CB5e0",
-      defaultDeliveryProviderContractAddress: "0x60a86b97a7596eBFd25fb769053894ed0D9A8366",
+      defaultDeliveryProviderContractAddress: testnetDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: "0xa5B7D85a8f27dd7907dc8FdC21FA5657D5E2F901",
       mockIntegrationAddress: "0x3bF0c43d88541BBCF92bE508ec41e540FbF28C56",
       rpcUrl: TESTNET_RPCS.moonbeam || "",
@@ -164,6 +168,8 @@ export const testnetEnv: Environment = {
   guardianRpcs: ["https://wormhole-v2-testnet-api.certus.one"],
 };
 
+export const mainnetDefaultDeliveryProviderContractAddress =
+  "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81";
 export const mainnetEnv: Environment = {
   network: "MAINNET",
   chainInfos: [
@@ -175,7 +181,7 @@ export const mainnetEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 2000,
       relayerContractAddress: "0x27428DD2d3DD32A4D7f7C497eAaa23130d894911",
-      defaultDeliveryProviderContractAddress: "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81",
+      defaultDeliveryProviderContractAddress: mainnetDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: CONTRACTS.MAINNET.ethereum.core,
       mockIntegrationAddress: "",
       rpcUrl: MAINNET_RPCS.ethereum || "",
@@ -188,7 +194,7 @@ export const mainnetEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 244,
       relayerContractAddress: "0x27428DD2d3DD32A4D7f7C497eAaa23130d894911",
-      defaultDeliveryProviderContractAddress: "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81",
+      defaultDeliveryProviderContractAddress: mainnetDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: CONTRACTS.MAINNET.bsc.core,
       mockIntegrationAddress: "",
       rpcUrl: MAINNET_RPCS.bsc || "",
@@ -201,7 +207,7 @@ export const mainnetEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 0.73,
       relayerContractAddress: "0x27428DD2d3DD32A4D7f7C497eAaa23130d894911",
-      defaultDeliveryProviderContractAddress: "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81",
+      defaultDeliveryProviderContractAddress: mainnetDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: CONTRACTS.MAINNET.polygon.core,
       mockIntegrationAddress: "",
       rpcUrl: MAINNET_RPCS.polygon || "",
@@ -214,7 +220,7 @@ export const mainnetEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 12,
       relayerContractAddress: "0x27428DD2d3DD32A4D7f7C497eAaa23130d894911",
-      defaultDeliveryProviderContractAddress: "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81",
+      defaultDeliveryProviderContractAddress: mainnetDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: CONTRACTS.MAINNET.avalanche.core,
       mockIntegrationAddress: "",
       rpcUrl: MAINNET_RPCS.avalanche || "",
@@ -227,7 +233,7 @@ export const mainnetEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 0.27,
       relayerContractAddress: "0x27428DD2d3DD32A4D7f7C497eAaa23130d894911",
-      defaultDeliveryProviderContractAddress: "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81",
+      defaultDeliveryProviderContractAddress: mainnetDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: CONTRACTS.MAINNET.fantom.core,
       mockIntegrationAddress: "",
       rpcUrl: MAINNET_RPCS.fantom || "",
@@ -241,7 +247,7 @@ export const mainnetEnv: Environment = {
     //   nativeCurrencyUsdPrice: 0.06,
     //   relayerContractAddress: "0x27428DD2d3DD32A4D7f7C497eAaa23130d894911",
     //   defaultDeliveryProviderContractAddress:
-    //     "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81",
+    //     mainnetDefaultDeliveryProviderContractAddress,
     //   coreBridgeAddress: "",
     //   mockIntegrationAddress: "",
     //   rpcUrl: MAINNET_RPCS.ethereum || "",
@@ -254,7 +260,7 @@ export const mainnetEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 0.16,
       relayerContractAddress: "0x27428DD2d3DD32A4D7f7C497eAaa23130d894911",
-      defaultDeliveryProviderContractAddress: "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81",
+      defaultDeliveryProviderContractAddress: mainnetDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: CONTRACTS.MAINNET.klaytn.core,
       mockIntegrationAddress: "",
       rpcUrl: MAINNET_RPCS.klaytn || "",
@@ -267,7 +273,7 @@ export const mainnetEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 0.49,
       relayerContractAddress: "0x27428DD2d3DD32A4D7f7C497eAaa23130d894911",
-      defaultDeliveryProviderContractAddress: "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81",
+      defaultDeliveryProviderContractAddress: mainnetDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: CONTRACTS.MAINNET.celo.core,
       mockIntegrationAddress: "",
       rpcUrl: MAINNET_RPCS.celo || "",
@@ -280,7 +286,7 @@ export const mainnetEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 0.24,
       relayerContractAddress: "0x27428DD2d3DD32A4D7f7C497eAaa23130d894911",
-      defaultDeliveryProviderContractAddress: "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81",
+      defaultDeliveryProviderContractAddress: mainnetDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: CONTRACTS.MAINNET.moonbeam.core,
       mockIntegrationAddress: "",
       rpcUrl: MAINNET_RPCS.moonbeam || "",
@@ -293,7 +299,7 @@ export const mainnetEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 2000,
       relayerContractAddress: "0x27428DD2d3DD32A4D7f7C497eAaa23130d894911",
-      defaultDeliveryProviderContractAddress: "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81",
+      defaultDeliveryProviderContractAddress: mainnetDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: CONTRACTS.MAINNET.arbitrum.core,
       mockIntegrationAddress: "",
       rpcUrl: MAINNET_RPCS.arbitrum || "",
@@ -306,7 +312,7 @@ export const mainnetEnv: Environment = {
       nativeCurrencyDecimals: 18,
       nativeCurrencyUsdPrice: 2000,
       relayerContractAddress: "0x27428DD2d3DD32A4D7f7C497eAaa23130d894911",
-      defaultDeliveryProviderContractAddress: "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81",
+      defaultDeliveryProviderContractAddress: mainnetDefaultDeliveryProviderContractAddress,
       coreBridgeAddress: CONTRACTS.MAINNET.optimism.core,
       mockIntegrationAddress: "",
       rpcUrl: MAINNET_RPCS.optimism || "",
